Move Dataset Details component to Dataset information, minor formatting cleanup in analysis router
