+++ client/src/entry/analysis/AnalysisRouter.js
@@ -17,40 +17,40 @@ import decodeUriComponent from "decode-uri-component";
 import Router from "layout/router";
 import ToolForm from "components/Tool/ToolForm";
 import FormGeneric from "components/Form/FormGeneric";
-import Sharing from "components/Sharing/Sharing.vue";
-import UserPreferences from "components/User/UserPreferences.vue";
-import DatasetList from "components/Dataset/DatasetList.vue";
+import Sharing from "components/Sharing/Sharing";
+import UserPreferences from "components/User/UserPreferences";
+import DatasetList from "components/Dataset/DatasetList";
 import { getUserPreferencesModel } from "components/User/UserPreferencesModel";
-import CustomBuilds from "components/User/CustomBuilds.vue";
+import CustomBuilds from "components/User/CustomBuilds";
 import Tours from "mvc/tours";
 import GridView from "mvc/grid/grid-view";
 import GridShared from "mvc/grid/grid-shared";
-import JobDetails from "components/JobInformation/JobDetails.vue";
-import WorkflowImport from "components/Workflow/WorkflowImport.vue";
-import TrsImport from "components/Workflow/TrsImport.vue";
-import TrsSearch from "components/Workflow/TrsSearch.vue";
-import InteractiveTools from "components/InteractiveTools/InteractiveTools.vue";
-import WorkflowList from "components/Workflow/WorkflowList.vue";
-import CollectionEditView from "components/Collections/common/CollectionEditView.vue";
-import HistoryImport from "components/HistoryImport.vue";
+import JobDetails from "components/JobInformation/JobDetails";
+import WorkflowImport from "components/Workflow/WorkflowImport";
+import TrsImport from "components/Workflow/TrsImport";
+import TrsSearch from "components/Workflow/TrsSearch";
+import InteractiveTools from "components/InteractiveTools/InteractiveTools";
+import WorkflowList from "components/Workflow/WorkflowList";
+import CollectionEditView from "components/Collections/common/CollectionEditView";
+import HistoryImport from "components/HistoryImport";
 import { HistoryExport } from "components/HistoryExport/index";
-import HistoryView from "components/HistoryView.vue";
-import WorkflowInvocationReport from "components/Workflow/InvocationReport.vue";
-import WorkflowRun from "components/Workflow/Run/WorkflowRun.vue";
+import HistoryView from "components/HistoryView";
+import WorkflowInvocationReport from "components/Workflow/InvocationReport";
+import WorkflowRun from "components/Workflow/Run/WorkflowRun";
 import UserInvocations from "components/Workflow/UserInvocations";
-import ToolsView from "components/ToolsView/ToolsView.vue";
-import ToolsJson from "components/ToolsView/ToolsSchemaJson/ToolsJson.vue";
+import ToolsView from "components/ToolsView/ToolsView";
+import ToolsJson from "components/ToolsView/ToolsSchemaJson/ToolsJson";
 import HistoryList from "mvc/history/history-list";
 import VisualizationsList from "components/Visualizations/Index";
 import QueryStringParsing from "utils/query-string-parsing";
 import DatasetError from "components/DatasetInformation/DatasetError";
 import DatasetAttributes from "components/DatasetInformation/DatasetAttributes";
-import Citations from "components/Citation/Citations.vue";
-import DisplayStructure from "components/DisplayStructured.vue";
+import Citations from "components/Citation/Citations";
+import DisplayStructure from "components/DisplayStructured";
 import { CloudAuth } from "components/User/CloudAuth";
 import { ExternalIdentities } from "components/User/ExternalIdentities";
-import Confirmation from "components/login/Confirmation.vue";
-import DatasetDetails from "components/Details/DatasetDetails.vue";
+import Confirmation from "components/login/Confirmation";
+import DatasetDetails from "components/DatasetInformation/DatasetDetails";
 import Libraries from "components/Libraries";
 import { mountVueComponent } from "utils/mountVueComponent";
